modules/wp_analyzer/config_audit.py
# Module for In-depth WordPress Configuration Audits
import requests
import re
from bs4 import BeautifulSoup
from core.utils import make_request # Assuming a utility for requests exists

def analyze_configuration(state, config, target_url):
    """
    Performs checks on various WordPress configuration settings for security best practices.
    Currently implements a heuristic check for the default DB prefix.
    Other checks (permissions, keys, htaccess) remain placeholders due to remote limitations.
    Updates the state with findings.
    """
    module_key = "wp_analyzer"
    findings_key = "configuration_audit"
    findings = state.get_specific_finding(module_key, findings_key, {
        "status": "Running",
        "details": "Performing configuration audit (DB prefix heuristic).",
        "db_prefix_check": {"status": "Running", "prefix_hint": None, "is_default_hint": None},
        "security_keys_check": {"status": "Not Implemented", "details": "Cannot check without wp-config.php access"},
        "file_permissions_check": {"status": "Not Implemented", "details": "Cannot check without server access/info leak"},
        "htaccess_check": {"status": "Not Implemented", "details": "Cannot check without server access/info leak"}
    })

    print("    Attempting to infer DB prefix via HTML source heuristics...")
    prefix_hint = None
    is_default_hint = None

    try:
        response = make_request(target_url, config, method="GET")
        if response and response.status_code == 200:
            soup = BeautifulSoup(response.text, 'lxml')
            # Heuristic: Look for input fields that might reveal the prefix (e.g., search forms sometimes use it)
            # This is unreliable but a common technique attempted.
            # Example: <input type="hidden" name="wp_posts_fields[]" value="post_title"> - unlikely but possible
            # More common might be in AJAX requests or specific plugin outputs, harder to generalize.
            # Let's check for a basic search form input name 's' which is default WP.
            search_input = soup.find('input', attrs={'name': 's'})
            if search_input:
                 # This doesn't directly reveal prefix, but confirms default search is likely used.
                 # A better heuristic might be needed, e.g., checking specific plugin outputs if known.
                 print("      [i] Found default search input ('s'). No direct prefix info.")
                 # We could try checking for common backup file names that include prefixes, but file_exposure does that.

            # Placeholder for a more advanced heuristic if developed later.
            # For now, we can't reliably detect the prefix remotely without info disclosure.
            findings["db_prefix_check"]["status"] = "Completed"
            findings["db_prefix_check"]["prefix_hint"] = "Undetermined"
            findings["db_prefix_check"]["is_default_hint"] = None
            findings["details"] = "DB prefix could not be reliably inferred from remote checks. Other checks not implemented."
            print("      [-] Could not reliably infer DB prefix remotely.")

        else:
            findings["db_prefix_check"]["status"] = "Error"
            findings["details"] = f"Failed to fetch homepage HTML (Status: {response.status_code if response else 'N/A'}) for DB prefix check."
            print(f"      [-] Failed to fetch homepage for DB prefix check.")

    except Exception as e:
        print(f"      [-] Error during DB prefix check: {e}")
        findings["db_prefix_check"]["status"] = "Error"
        findings["details"] = f"Error during DB prefix check: {e}"


    # Add general note about default prefix if we couldn't determine otherwise
    if findings["db_prefix_check"]["prefix_hint"] == "Undetermined":
         state.add_remediation_suggestion("db_prefix_default_check", {
            "source": "WP Analyzer",
            "description": "Could not determine the database table prefix remotely. Using the default 'wp_' prefix is common but less secure.",
            "severity": "Low",
            "remediation": "Consider changing the default 'wp_' database table prefix during installation or using a security plugin to change it post-installation. This makes some automated SQL injection attacks slightly harder."
        })

    findings["status"] = "Completed" # Mark module as done
    state.update_specific_finding(module_key, findings_key, findings)
    print(f"    [+] Configuration audit finished. Details: {findings['details']}")
